`retry_delay: float`

+++ mqclient/broker_client_interface.py
@@ -122,7 +122,7 @@ class Pub(RawQueue):
         self,
         msg: bytes,
         retries: int,
-        retry_delay: int,
+        retry_delay: float,
     ) -> None:
         """Send a message on a queue."""
         raise NotImplementedError()
@@ -146,7 +146,7 @@ class Sub(RawQueue):
         self,
         timeout_millis: Optional[int],
         retries: int,
-        retry_delay: int,
+        retry_delay: float,
     ) -> Optional[Message]:
         """Get a single message from a queue."""
         raise NotImplementedError()
@@ -155,7 +155,7 @@ class Sub(RawQueue):
         self,
         msg: Message,
         retries: int,
-        retry_delay: int,
+        retry_delay: float,
     ) -> None:
         """Ack a message from the queue."""
         raise NotImplementedError()
@@ -164,7 +164,7 @@ class Sub(RawQueue):
         self,
         msg: Message,
         retries: int,
-        retry_delay: int,
+        retry_delay: float,
     ) -> None:
         """Reject (nack) a message from the queue."""
         raise NotImplementedError()
@@ -174,7 +174,7 @@ class Sub(RawQueue):
         timeout: int,
         propagate_error: bool,
         retries: int,
-        retry_delay: int,
+        retry_delay: float,
     ) -> AsyncGenerator[Optional[Message], None]:
         """Yield Messages.
